requirements.txt
numpy==1.19.1
pathlib==1.0.1
Pillow==7.2.0
PyQt5==5.15.0
selenium==3.141.0
webdriver-manager==3.2.1